+++ test/unit.test.js
@@ -39,7 +39,8 @@ const nodeAndBrowser = [
 
 const nodeOnly = [
     '../test_tmp/unit/plugin.test.js',
-    '../test_tmp/unit/no-validate.test.js'
+    '../test_tmp/unit/no-validate.test.js',
+    '../test_tmp/unit/ajv-validate.test.js'
 ];
 
 const typings = [